+++ packages/vsce/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "cics-extension-for-zowe" extension will be documented in this file.
 
+## Recent Changes
+
+ - Enhancement: Lack of logging in the CICS VS Code extension. [#250](https://github.com/zowe/cics-for-zowe-client/issues/250)
+
 ## `3.4.0`
 
 - Enhancement: Use LTPA tokens to allow CMCI "sessions" [#217](https://github.com/zowe/cics-for-zowe-client/issues/217)